feat: allow using conventional-commits 🎨
We already have a Commitlint ruleset which enforces conventional-commits strategy, so this PR is mostly to update our Git guidelines to match recommendations from conevntional-commits.

The general benefit of allowing conventional-commits is that

- it is considered an industry standard (or is at least very popular)
- there are tools to automate many tasks "for free" if we simply follow conventional-commits, like automated versioning management, changelog auto-generation, publishing to npm and many other things

This is a similar "debate" as whether or not people want/should use Prettier. One might not like how the code looks but it sure looks consistent across all projects which use it.

So let's just use it! 💪